[REFACTOR]: RedisService infra 패키지로 이동
