+++ src/main/java/project/infra/redis/application/RedisService.java
@@ -1,4 +1,4 @@
-package project.common.service;
+package project.infra.redis.application;
 
 public interface RedisService {
