+++ src/test/java/project/common/config/TestRedisConfig.java
@@ -3,7 +3,7 @@ package project.common.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
-import project.common.service.RedisService;
+import project.infra.redis.application.RedisService;
 
 @Configuration
 public class TestRedisConfig {
